Restrict sound manipulation to owner

+++ src/lib/server/bot/guild-state.ts
@@ -1,4 +1,5 @@
 import type { GuildDataSoundPatch, GuildDataSoundSnippet } from '$lib/snippets';
+import { error } from '@sveltejs/kit';
 import type { Message } from 'discord.js';
 import { and, eq } from 'drizzle-orm';
 import { server } from '..';
@@ -41,8 +42,7 @@ export class GuildData {
 		);
 	}
 
-	async createSound(options: CreateSoundOptions): Promise<GuildDataSoundSnippet> {
-		const { userId, name, request } = options;
+	async createSound({ userId, name, request }: CreateSoundOptions): Promise<GuildDataSoundSnippet> {
 		const asset = await server.assetManager.uploadAsset(userId, request);
 
 		const sound: table.Sound = {
@@ -56,43 +56,57 @@ export class GuildData {
 		await db.insert(table.sound).values(sound);
 		return {
 			id: sound.id,
+			mediaPath: server.assetManager.resolveAssetPath(asset.path),
 			name: sound.name,
 			keywords: sound.keywords,
-			mediaPath: server.assetManager.resolveAssetPath(asset.path)
+			createdBy: userId
 		};
 	}
 
-	async patchSound(id: string, partialSound: GuildDataSoundPatch): Promise<GuildDataSoundSnippet> {
+	async patchSound({ userId, id, patch }: PatchSoundOptions): Promise<GuildDataSoundSnippet> {
 		const [original = undefined] = await db
 			.select()
 			.from(table.sound)
 			.where(and(eq(table.sound.id, id), eq(table.sound.guildId, this.guildState.id)))
 			.leftJoin(table.asset, eq(table.sound.assetId, table.asset.id));
 
-		if (!original) {
-			throw new Error(`Invalid sound ID ${id}`);
+		if (!original || !original.asset) {
+			throw error(400, `Invalid sound ID ${id}`);
 		}
 
-		await db.update(table.sound).set(partialSound).where(eq(table.sound.id, id));
+		if (original.asset.createdBy !== userId) {
+			throw error(403, { message: 'Sound is owned by a different user' });
+		}
+
+		await db.update(table.sound).set(patch).where(eq(table.sound.id, id));
 
 		return {
 			id: original.sound.id,
 			name: original.sound.name,
 			keywords: original.sound.keywords,
-			mediaPath: server.assetManager.resolveAssetPath(original.asset!.path),
-			...partialSound
+			createdBy: original.asset.createdBy,
+			mediaPath: server.assetManager.resolveAssetPath(original.asset.path),
+			...patch
 		};
 	}
 
-	async deleteSound(id: string) {
+	async deleteSound({ userId, id }: DeleteSoundOptions) {
 		const [entry = undefined] = await db
-			.select({ assetId: table.asset.id, assetPath: table.asset.path })
+			.select({
+				assetId: table.asset.id,
+				assetPath: table.asset.path,
+				assetOwner: table.asset.createdBy
+			})
 			.from(table.sound)
 			.where(and(eq(table.sound.guildId, this.guildState.id), eq(table.sound.id, id)))
 			.leftJoin(table.asset, eq(table.sound.assetId, table.asset.id));
 
 		if (!entry) {
-			throw new Error(`Invalid sound ID ${id}`);
+			throw error(400, `Invalid sound ID ${id}`);
+		}
+
+		if (entry.assetOwner !== userId) {
+			throw error(403, { message: 'Sound is owned by a different user' });
 		}
 
 		await server.assetManager.deleteAsset({ id: entry.assetId!, path: entry.assetPath! });
@@ -103,4 +117,15 @@ interface CreateSoundOptions {
 	userId: string;
 	name: string;
 	request: Request;
+}
+
+interface PatchSoundOptions {
+	userId: string;
+	id: string;
+	patch: GuildDataSoundPatch;
+}
+
+interface DeleteSoundOptions {
+	userId: string;
+	id: string;
 }
